# request-counter-api/src/main/java/io/swagger/api/RequestsCounterApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.22).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.CounterMessage;
import io.swagger.model.ErrorModel;
import io.swagger.model.Request;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-10-08T08:00:53.858Z[GMT]")
@Api(value = "requests-counter", description = "the requests-counter API")
public interface RequestsCounterApi {

    @ApiOperation(value = "Increments counter", nickname = "addRequest", notes = "Increments counter and returns new requests count", response = CounterMessage.class, tags={ "Requests", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Message returned", response = CounterMessage.class),
        @ApiResponse(code = 200, message = "error payload", response = ErrorModel.class) })
    @RequestMapping(value = "/requests-counter",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<CounterMessage> addRequest(@ApiParam(value = ""  )  @Valid @RequestBody Request body
);


    @ApiOperation(value = "Returns counter value", nickname = "getNumberOfRequests", notes = "Returns number of requests made by users", response = CounterMessage.class, tags={ "Requests", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Message returned", response = CounterMessage.class),
        @ApiResponse(code = 200, message = "error payload", response = ErrorModel.class) })
    @RequestMapping(value = "/requests-counter",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<CounterMessage> getNumberOfRequests();

}